Moved UserTweetFilter to new Helpers package, created IntentHelp
